Renamed BookGenreComboBox => BookGenre & moved it
